reorganize lib tests
