+++ .gitignore
@@ -15,3 +15,4 @@ stashed
 /playwright-report/
 /blob-report/
 /playwright/.cache/
+/test/tmp/**/*/
