+++ test/lib/bin.test.js
@@ -1,6 +1,6 @@
 /* globals describe it */
 import * as assert from 'assert';
-import fn from '../src/lib/lib/bin.js';
+import fn from '../../src/lib/lib/bin.js';
 
 const name = 'bin';
