+++ test/lib/calcExtents.test.js
@@ -1,6 +1,6 @@
 /* globals describe it */
 import * as assert from 'assert';
-import fn from '../src/lib/lib/calcExtents.js';
+import fn from '../../src/lib/lib/calcExtents.js';
 
 const name = 'calcExtents';
