+++ test/lib/calcUniques.test.js
@@ -1,6 +1,6 @@
 /* globals describe it */
 import * as assert from 'assert';
-import fn from '../src/lib/lib/calcUniques.js';
+import fn from '../../src/lib/lib/calcUniques.js';
 
 const name = 'calcUniques';
