+++ test/lib/filterObject.test.js
@@ -1,6 +1,6 @@
 /* globals describe it */
 import * as assert from 'assert';
-import fn from '../src/lib/utils/filterObject.js';
+import fn from '../../src/lib/utils/filterObject.js';
 
 const name = 'filterObject';
