+++ test/lib/groupLonger.test.js
@@ -1,6 +1,6 @@
 /* globals describe it */
 import * as assert from 'assert';
-import fn from '../src/lib/lib/groupLonger.js';
+import fn from '../../src/lib/lib/groupLonger.js';
 
 const name = 'groupLonger';
