+++ test/lib/isOrdinalDomain.test.js
@@ -2,7 +2,7 @@
 import * as assert from 'assert';
 import * as d3Scale from 'd3-scale';
 
-import fn from '../src/lib/helpers/isOrdinalDomain.js';
+import fn from '../../src/lib/helpers/isOrdinalDomain.js';
 
 const name = 'isOrdinalScale';
